Extract PositionSlider widget
